requirements.txt
torch==2.3.0
sentencepiece
tiktoken
accelerate
opencv-python
Pillow
transformers
numpy

